+++ test/dlc_tests/ec2/test_efa.py
@@ -90,23 +90,6 @@ def test_pytorch_efa(
     is_pr_context() and not is_efa_dedicated(),
     reason="Skip EFA test in PR context unless explicitly enabled",
 )
-def test_pytorch_transformerengine(
-    pytorch_training, efa_ec2_instances, efa_ec2_connections, ec2_instance_type, region, gpu_only
-):
-    pass
-
-
-@pytest.mark.processor("gpu")
-@pytest.mark.model("N/A")
-@pytest.mark.integration("efa")
-@pytest.mark.usefixtures("sagemaker")
-@pytest.mark.allow_p4de_use
-@pytest.mark.multinode(2)
-@pytest.mark.parametrize("ec2_instance_type,region", EC2_EFA_GPU_INSTANCE_TYPE_AND_REGION)
-@pytest.mark.skipif(
-    is_pr_context() and not is_efa_dedicated(),
-    reason="Skip EFA test in PR context unless explicitly enabled",
-)
 def test_efa_tensorflow(
     tensorflow_training, efa_ec2_instances, efa_ec2_connections, ec2_instance_type, region, gpu_only
 ):
